ref(program): <- factor out BTC sub mat in that
